+++ src/orquestra/sdk/_base/abc.py
@@ -22,10 +22,12 @@ from orquestra.sdk.schema.ir import TaskInvocationId, WorkflowDef
 from orquestra.sdk.schema.local_database import StoredWorkflowRun
 from orquestra.sdk.schema.responses import WorkflowResult
 from orquestra.sdk.schema.workflow_run import (
+    ProjectId,
     State,
     WorkflowRun,
     WorkflowRunId,
     WorkflowRunMinimal,
+    WorkspaceId,
 )
 
 
@@ -159,6 +161,8 @@ class RuntimeInterface(ABC):
         limit: t.Optional[int] = None,
         max_age: t.Optional[timedelta] = None,
         state: t.Optional[t.Union[State, t.List[State]]] = None,
+        workspace: t.Optional[WorkspaceId] = None,
+        project: t.Optional[ProjectId] = None
     ) -> t.Sequence[WorkflowRunMinimal]:
         """
         List the workflow runs, with some filters
@@ -167,6 +171,8 @@ class RuntimeInterface(ABC):
             limit: Restrict the number of runs to return, prioritising the most recent.
             max_age: Only return runs younger than the specified maximum age.
             state: Only return runs of runs with the specified status.
+            workspace: Only return runs from the specified workspace. Supported only
+                on CE.
         Returns:
                 A list of the workflow runs
         """